fix(content): fix svelte 5 on-mount (#206)
`$effect` runs after the component gets mounted and rendered, but `pageTitle` isn't reactive. Thus, the page won't re-render with the new value.
For this logic, it's better to use `$effert.pre` - it runs before mounting and the first rendering. But as the example concerns the on-mount event, it's better to make `pageTitle` reactive.